ENH : change test pixeltype from int to float

+++ Testing/Code/IO/otbImageFileReaderRADComplexFloatExtract.cxx
@@ -46,8 +46,8 @@ int otbImageFileReaderRADComplexFloatExtract(int argc, char* argv[])
 
   typedef otb::Image<InputPixelType,  Dimension> InputImageType;
   typedef otb::Image<OutputPixelType, Dimension> OutputCplxImageType;
-  typedef otb::Image<int, Dimension>             OutputScalarImageType;
-  typedef otb::VectorImage<int, Dimension>       OutputImageType;
+  typedef otb::Image<float, Dimension>             OutputScalarImageType;
+  typedef otb::VectorImage<float, Dimension>       OutputImageType;
   typedef otb::ImageList<OutputScalarImageType>  ImageListType;
 
   typedef otb::ImageFileReader<InputImageType>  ReaderType;
